Non-Adaptive-Source/requirements.txt
albumentations==1.1.0
boto3==1.20.20
hydra-core==1.1.1
hydra-colorlog==1.1.0
hydra-optuna-sweeper==1.1.0
lightning_bolts==0.5.0
matplotlib==3.4.2
numpy==1.17.5
omegaconf==2.1.1
opencv_python_headless==4.5.4.60
packaging==20.9
Pillow==9.0.0
pycocotools==2.0
pytest==6.2.5
python-dotenv==0.19.2
pytorch_lightning==1.5.4
rich==11.0.0
scikit_learn==1.0.2
seaborn==0.11.2
setuptools==58.0.4
sh==1.14.2
torch==1.10.0
torchmetrics==0.9.3
torchvision==0.11.1
tqdm==4.62.3
whos_there==0.1.1
wandb
git+https://github.com/philferriere/cocoapi.git#egg=pycocotools&subdirectory=PythonAPI # pycocotools
